cofiguered EsLint & Prettier

+++ migrations/sqls/20220926094832-orders-table-up.sql
@@ -1,1 +1,1 @@
-CREATE TABLE orders (id SERIAL PRIMARY KEY, order_status BOOLEAN, user_id bigint REFERENCES users(id));
+CREATE TABLE orders (id SERIAL PRIMARY KEY, order_status BOOLEAN, user_id int REFERENCES users(id));
